# Conflicts:
#	.idea/gradle.xml
#	.idea/misc.xml
#	Java-OpenMobility
#	app/app.iml
#	app/src/main/java/br/org/funcate/terramobile/controller/activity/MainActivity.java
#	app/src/main/java/br/org/funcate/terramobile/model/tilesource/AppGeoPackageService.java
#	app/src/main/res/values/georesource.xml
#	dynamicforms/dynamicforms.iml

+++ app/src/main/java/br/org/funcate/terramobile/controller/activity/MapFragment.java
@@ -198,8 +198,8 @@ public class MapFragment extends Fragment implements OpenStreetMapConstants{
     }
 
     public void drawCross(ImageView drawingImageView) {
-        Bitmap bitmap = Bitmap.createBitmap((int) getActivity().getWindowManager()
-                .getDefaultDisplay().getWidth(), (int) getActivity().getWindowManager()
+        Bitmap bitmap = Bitmap.createBitmap(getActivity().getWindowManager()
+                .getDefaultDisplay().getWidth(), getActivity().getWindowManager()
                 .getDefaultDisplay().getHeight(), Bitmap.Config.ARGB_8888);
 
         Canvas canvas = new Canvas(bitmap);
